# docs/openapi/aether-v0.yaml
openapi: 3.1.0
info:
  title: Aether /api/v0 API
  version: 0.1.0
  contact:
    name: Andres Bott
    url: https://github.com/andresbott/aether
  description: >
    Internal, admin-only server-management API — collections/libraries,
    scanning, tasks/scheduling, and other administration concerns. This is
    NOT the music-client API: browsing, playback, playlists, album/artist
    navigation and search live exclusively under `/rest`, which must stay
    OpenSubsonic-compliant and has no OpenAPI document of its own (a Subsonic
    client does not consume one).


    **Scope.** This document describes the full `/api/v0` surface as mounted
    in `app/router/api_v0.go`: the metadata editor (pictures, raw tags,
    folder/track browsing, identify and identify-album, artist images),
    artists, users, tasks, tokens, radiobrowser, auth, libraries, and
    health/version/me.
servers:
  - url: /api/v0
    description: >
      Every path below is mount-relative. The base is modelled exclusively
      through this `servers` entry, so bumping the version prefix — this is
      `/api/v0` while the API is pre-1.0 — stays a one-line change, with no
      path or operationId edits.
...
security:
  - sessionCookie: []
paths:
...
  /metadata/identify:
    post:
      operationId: identifyTracks
      summary: Resolve a track selection to MusicBrainz recording candidates by acoustic fingerprint.
      description: >
        Fingerprints each file (fpcalc) and looks it up against AcoustID,
        sequentially — fingerprinting is CPU-bound and the AcoustID client
        is rate-limited, so concurrency buys nothing. A per-file failure (an
        unreadable file, a path outside the library, an upstream lookup
        failure) is reported on that file's own result row, verbatim, and
        never fails the whole request; the request itself only fails
        validation (a bad selection) or answers 503 when no identification
        service is configured — see `getMetadataCapabilities`.
      tags: [metadata-identify]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentifyRequest' }
      responses:
        '200':
          description: One result per requested path, in request order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdentifyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /metadata/identify-album:
    post:
      operationId: identifyAlbum
      summary: Map a multi-file selection onto a single best-matching MusicBrainz release.
      description: >
        Fingerprints every file, unions every release any candidate
        mentions into ranked options, enriches the best few (capped at 8,
        since MusicBrainz is rate-limited) with their full tracklist, and
        places files the fingerprint missed onto the enriched tracklist by
        duration/title/track-number hints. A partial failure (one bad file
        among many) is never a request failure — it is reported in `errors`
        and the option list is still built from the rest. The request
        itself fails (429/502) only when EVERY file failed and at least one
        failure came from the AcoustID service itself (an outage), rather
        than from the files; it answers 503 when no album-identification
        service is configured — see `getMetadataCapabilities`.
      tags: [metadata-identify]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentifyAlbumRequest' }
      responses:
        '200':
          description: >
            Candidate releases, best first, each carrying one assignment
            per input file, plus the files this identification did not
            cover.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IdentifyAlbumResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '502': { $ref: '#/components/responses/UpstreamError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
...
  /users:
    get:
      operationId: listUsers
      summary: List native user accounts.
      description: >
        Admin. Native-mode only: the whole users group is unmounted unless
        the server's auth method is "native" — in proxy mode users are
        provisioned from the reverse-proxy identity provider on first sight
        instead (see `GET /me`'s `features.userManagement`, which reports
        this group's availability). Unpaginated: the store is asked for at
        most 200 rows, comfortably covering a self-hosted deployment, so
        `total` and the returned array's length agree in practice.
      tags: [users]
      responses:
        '200':
          description: Every native user account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createUser
      summary: Create a native user account.
      description: >
        Admin. Native-mode only (see `listUsers`). `enabled` omitted
        defaults to true; `role` omitted defaults to "user". `login` is
        rejected with 422 when it is shaped like a token id (10 lowercase
        letters/digits) — that namespace is reserved for `/rest` token
        authentication (`u`+`t`+`s`/`u`+`p`), where the virtual username IS a
        token id, so a colliding login would be unreachable by its own name.
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: >
            Created. The response includes the store-generated id, read back
            after the insert (CreateUser does not return it directly) — the
            same id updateUser/deleteUser address.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: _aether_login
      description: >
        Native-mode session cookie (see docs/agents/authentication.md).
        `/api/v0` defaults every route to admin-only, so a valid, enabled
        admin session is required for the whole metadata surface. This is the
        only credential `/api/v0` actually validates: no route on this
        surface reads an `Authorization` header, and Aether's PATs
        (bearer-shaped credentials) are verified exclusively on `/rest` (via
        `apiKey`, or `u`+`t`+`s` / `u`+`p`), never here.


        In **proxy-header mode**, this cookie plays no role either — identity
        instead comes from reverse-proxy-injected headers (`Remote-User` /
        `Remote-Groups`), validated by `headerGuard`. That trust is
        established between the proxy and Aether, not chosen by the API
        caller, so it has no client-facing credential to model as an OpenAPI
        security scheme; it is intentionally left undeclared here rather than
        represented by a scheme nothing in the protocol actually sends.
...
  schemas:
    Problem:
      type: object
      description: >
        RFC 9457 Problem Details. `type` is a stable, never-fetched URI whose
        last path segment mirrors the old ad-hoc "code" string.
      properties:
        type:
          type: string
          format: uri
          description: Never dereferenced; only the last path segment is meant to be read.
          example: "https://aether.local/probs/validation_error"
        title:
          type: string
          example: "Validation error"
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          description: The request path.
      required: [type, title, status]
    FieldError:
      type: object
      properties:
        pointer:
          type: string
          description: JSON Pointer (RFC 6901) to the failing field, e.g. "/paths".
        detail:
          type: string
      required: [pointer, detail]
    ValidationProblem:
      description: >
        A Problem for a well-formed but invalid (422) request, itemising
        which fields failed and why. Every 422 on this surface is produced by
        the same code path and carries exactly one field error today, but
        `errors` is not capped at one.
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items: { $ref: '#/components/schemas/FieldError' }
...
    PictureSelection:
      type: object
      description: >
        The library and the selected track paths (library-relative), carried
        in the body rather than the URL so a large multi-disc selection can
        never overflow a header buffer — the production 431 this redesign
        fixes. The JSON request body is capped at 1 MiB; a larger body is
        rejected as a malformed-request 400.
      properties:
        library_id: { type: integer, format: int64, minimum: 0 }
        paths:
          type: array
          items: { type: string }
          minItems: 1
          maxItems: 50
          description: Selected track paths, library-relative.
      required: [library_id, paths]
...
    ArtistCredit:
      type: object
      description: >
        One credited artist: the credited-as name and its MusicBrainz id.
        Shared wire shape of `identify.go`'s `identifyArtistDTO` (a
        fingerprinted recording's artists) and `albumidentify.Artist` (an
        album option's, or a track assignment's, artists) — two distinct Go
        types with an identical `{name, mbid}` shape.
      properties:
        name: { type: string }
        mbid: { type: string }
      required: [name, mbid]
    IdentifyRelease:
      type: object
      description: >
        One release (album) a fingerprinted recording appears on.
        `track_number`/`disc_number` locate it on that release; `0` means
        unknown.
      properties:
        release_mbid: { type: string }
        release_group_mbid: { type: string }
        album: { type: string }
        year: { type: integer }
        track_number: { type: integer }
        disc_number: { type: integer }
      required:
        - release_mbid
        - release_group_mbid
        - album
        - year
        - track_number
        - disc_number
    IdentifyCandidate:
      type: object
      description: One MusicBrainz recording matched by acoustic fingerprint, best first.
      properties:
        score: { type: number }
        recording_mbid: { type: string }
        title: { type: string }
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistCredit' }
        releases:
          type: array
          items: { $ref: '#/components/schemas/IdentifyRelease' }
      required: [score, recording_mbid, title, artists, releases]
    IdentifyResult:
      type: object
      description: One requested path's identification outcome.
      properties:
        path: { type: string }
        candidates:
          type: array
          items: { $ref: '#/components/schemas/IdentifyCandidate' }
        error:
          type: string
          description: >
            Set instead of (an empty) `candidates` when this file could not
            be resolved (outside the library, unreadable, or the AcoustID
            lookup failed) — the raw error text, since this endpoint has no
            `albumidentify`-style fixed-sentence policy.
      required: [path, candidates]
    IdentifyRequest:
      description: >
        The same `{library_id, paths[]}` shape as the picture-selection
        endpoints; at least one path is required.
      allOf:
        - $ref: '#/components/schemas/PictureSelection'
    IdentifyResponse:
      type: object
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/IdentifyResult' }
      required: [results]
    AlbumTrackSlot:
      type: object
      description: One position on a release's tracklist, from MusicBrainz enrichment.
      properties:
        disc_number: { type: integer }
        track_number: { type: integer }
        title: { type: string }
        recording_mbid: { type: string }
        duration_seconds: { type: number }
      required: [disc_number, track_number, title, recording_mbid, duration_seconds]
    AlbumAssignment:
      type: object
      description: >
        One selected file's placement on one `AlbumOption`. `source`
        distinguishes how strong the placement is: `fingerprint` (AcoustID
        matched this file to a recording the release carries at this
        position), `inferred` (placed by duration/title/track-number
        against the release's tracklist instead) or `none` (no placement
        could be justified; only the option's album-level fields apply to
        this file).
      properties:
        path: { type: string }
        source:
          type: string
          enum: [fingerprint, inferred, none]
        title: { type: string }
        recording_mbid: { type: string }
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistCredit' }
          description: >
            Always an array, never `null`, even for a `none`/`inferred`
            placement with nothing to report — see `AlbumOption` for why
            this matters on the wire.
        disc_number: { type: integer }
        track_number: { type: integer }
        score: { type: number }
        error:
          type: string
          description: >
            Set, instead of a placement, when this file could not be
            fingerprinted/looked up at all — one of `albumidentify`'s three
            fixed reasons, never a raw error (unlike `IdentifyResult.error`).
      required:
        - path
        - source
        - title
        - recording_mbid
        - artists
        - disc_number
        - track_number
        - score
    AlbumOption:
      type: object
      description: >
        One candidate release for the whole selection, best first, with one
        `AlbumAssignment` per input file. **Every slice field
        (`artists`/`tracks`/`assignments`) is always an array, never
        `null`** — enforced by `internal/albumidentify`'s
        `TestResolveNeverEmitsNullSlicesInJSON`: `artists`/`tracks` start as
        `[]` at construction and are only ever replaced (never nulled) by
        `enrich`, and `assignments` gains one entry per input file whether
        or not `enrich` ran. `enriched` false means the MusicBrainz
        tracklist lookup did not happen (past `MaxEnrichedOptions`, 8) or
        failed: `track_count`/`disc_count`/`tracks` are then unknown
        (zero/empty) and no gap-fill against a tracklist was possible, so
        every assignment is either `fingerprint` or `none`.
      properties:
        release_mbid: { type: string }
        release_group_mbid: { type: string }
        album: { type: string }
        year: { type: integer }
        artists:
          type: array
          items: { $ref: '#/components/schemas/ArtistCredit' }
        track_count: { type: integer }
        disc_count: { type: integer }
        enriched: { type: boolean }
        matched_count:
          type: integer
          description: How many files a fingerprint (not inference) placed on this release.
        mean_score:
          type: number
          description: >
            Mean fingerprint score over `matched_count` files; unaffected
            by inferred placements.
        assignments:
          type: array
          items: { $ref: '#/components/schemas/AlbumAssignment' }
        tracks:
          type: array
          items: { $ref: '#/components/schemas/AlbumTrackSlot' }
      required:
        - release_mbid
        - release_group_mbid
        - album
        - year
        - artists
        - track_count
        - disc_count
        - enriched
        - matched_count
        - mean_score
        - assignments
        - tracks
    IdentifyPathError:
      type: object
      description: >
        One requested file `/metadata/identify-album` did not cover: either
        refused before identification (outside the library) or a file that
        reached the resolver but could not be fingerprinted/looked up.
        `error` is always one of `albumidentify`'s fixed, short sentences —
        never a raw error, since this is rendered verbatim in the dialog.
      properties:
        path: { type: string }
        error: { type: string }
      required: [path, error]
    IdentifyAlbumRequest:
      description: >
        The same `{library_id, paths[]}` shape as the picture-selection
        endpoints. Identifying "the album" needs more than one file to
        place: the handler requires **at least two** paths — a single path
        is better served by `/metadata/identify` — and answers 422
        (`ValidationProblem`, itemising `/paths`) for a shorter selection,
        the same as an empty selection on any other endpoint. This schema
        enforces that floor with the `minItems: 2` override below, which
        narrows the `minItems: 1` inherited from `PictureSelection`.
      allOf:
        - $ref: '#/components/schemas/PictureSelection'
        - type: object
          properties:
            paths:
              type: array
              items: { type: string }
              minItems: 2
    IdentifyAlbumResult:
      type: object
      properties:
        options:
          type: array
          items: { $ref: '#/components/schemas/AlbumOption' }
          description: >
            Candidate releases, best first. Always an array, never `null`
            — including the empty-selection case (every path was outside
            the library) and the no-release-matched-anything case.
        errors:
          type: array
          items: { $ref: '#/components/schemas/IdentifyPathError' }
      required: [options, errors]
...
    User:
      type: object
      description: >
        A native user account. Never carries a password or its hash — write
        one via `password` on CreateUserRequest/UpdateUserRequest, never
        read it back.
      properties:
        id:
          type: string
          format: uuid
          description: Stable id, minted once at creation; addresses updateUser/deleteUser.
        login:
          type: string
        enabled:
          type: boolean
        role:
          type: string
          enum: [admin, user]
          description: >
            The derived vertical, not the raw group list: "admin" means
            membership in the store's admin group, "user" means none.
      required: [id, login, enabled, role]
    UserList:
      type: object
      properties:
        users:
          type: array
          items: { $ref: '#/components/schemas/User' }
        total:
          type: integer
          description: >
            Total native user count, ignoring the server's internal 200-row
            page cap — in every deployment under that cap (every
            self-hosted server today) it equals `users`' length.
      required: [users, total]
    CreateUserRequest:
      type: object
      properties:
        login:
          type: string
          description: >
            Required (400 if blank). Rejected with 422 when it is shaped
            like a token id (10 lowercase letters/digits) — see
            `createUser`.
        password:
          type: string
          description: >
            Required (400 if blank). Longer than 72 bytes is rejected with
            422: bcrypt only hashes a password's first 72 bytes, so a longer
            input would otherwise silently truncate instead of being
            honoured in full.
        enabled:
          type: boolean
          default: true
        role:
          type: string
          enum: [admin, user]
          default: user
      required: [login, password]
...
  responses:
    BadRequest:
      description: Malformed request (bad JSON/multipart, missing/invalid scalar).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: The addressed resource does not exist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: >
        The request is well-formed but semantically invalid (over a size
        cap, an unknown type/slot, an empty selection); `errors` itemises
        the failing fields.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
    TooManyRequests:
      description: A third-party upstream service is rate-limiting Aether's requests.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UpstreamError:
      description: >
        A third-party upstream service call failed. The same upstream
        classification can also surface as 504 on a timeout (still this
        Problem shape) — see `GatewayTimeout`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    GatewayTimeout:
      description: A third-party upstream service timed out.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: No valid session (native mode) — the caller must authenticate.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted (admin required, or the user is disabled).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: The request conflicts with current server state.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServiceUnavailable:
      description: >
        The requested capability is not configured server-side (e.g.
        `not_configured` — no image-provider API key set). The generic
        Problem shape; see `type`/`detail` for the specific reason.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalServerError:
      description: >
        An unexpected server-side failure — a store or internal error that is
        not a not-found (e.g. a failed database read/write, a filesystem write
        that did not complete). The generic Problem shape; see `type`/`detail`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }